Final Update

+++ app/Http/Controllers/NewsController.php
@@ -5,23 +5,30 @@ namespace App\Http\Controllers;
 use App\News;
 use Illuminate\Http\Request;
 use phpDocumentor\Reflection\Types\Integer;
+use Illuminate\Support\Facades\DB;
+use function GuzzleHttp\Promise\all;
 
 class NewsController extends Controller
 {
 
     public function index()
     {
-        $news=News::all();
+        /**
+         * @var News $news
+         */
+        $news=News::orderBy('updated_at', 'DESC')->paginate(5);
         return view('newses.index', compact('news'));
     }
 
     public function create()
     {
+        $this->authorize('create', News::class);
         return view('newses.form');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', News::class);
         $data=$this->validated($request);
 
         $news=News::on()->create($data);
@@ -31,16 +38,19 @@ class NewsController extends Controller
 
     public function show(News $newse)
     {
-        return view('newses.show', compact('newse'));
+        $comments=$newse->comments()->orderBy('updated_at', 'DESC')->get();
+        return view('newses.show', compact('newse','comments'));
     }
 
     public function edit(News $newse)
     {
+        $this->authorize('update', $newse);
         return view('newses.form', compact('newse'));
     }
 
     public function update(Request $request, News $newse)
     {
+        $this->authorize('update', $newse);
         $data=$this->validated($request,$newse);
 
         $newse->update($data);
@@ -49,12 +59,13 @@ class NewsController extends Controller
 
     public function destroy(News $newse)
     {
+        $this->authorize('delete', $newse);
         $newse->delete();
         return redirect()->route('newses.index');
     }
 
     public function getNews(Request $request,int $id){
-        $news = News::where('category', $id)->get();
+        $news = News::where('category', $id)->orderBy('updated_at', 'DESC')->paginate(5);
         return view('newses.index', compact('news'));
     }
 
@@ -64,14 +75,11 @@ class NewsController extends Controller
             'description'=>'nullable',
             'imageURL'=>'nullable',
             'category'=>'category'
-            //Добавил custom валидацию для status по имени status, который зареган в App\AppServiceProvider
         ];
-        //Чтобы он не проверял самого себя по id
         if($news)
             $rules['title'] .=',title,' . $news->id;
 
         return $request->validate($rules, [
-            //Ключ(status) относится к имени валидации
             'category.category'=>'Category doesn\'t match'
         ]);
     }
